Add a CRC file wrapper library.
The goal is to add a CRC to protect the data section of a file.

The CRC is stored in the extended attribute of the file inode.

Therefore the CRC protecting the data is itself protected by the CRC
of the inode.

A function is added to check the file CRC against the one in the
metadata.

Another function allows to fix the file CRC.

Signed-off-by: Jean-Christophe Dubois <jcd@tribudubois.net>